Conflicts:
	localwiki/maps/urls.py
	localwiki/themes/sapling/assets/css/site.css

+++ localwiki/users/signals.py
@@ -1,6 +1,8 @@
 from django.contrib.auth.models import User
 from django.db.models.signals import post_save, pre_delete
 
+from south.models import MigrationHistory
+
 from models import UserProfile
 
 
@@ -10,14 +12,21 @@ def create_user_profile(sender, instance, created, raw, **kwargs):
         # being imported.
         return
     if created:
-        UserProfile.objects.create(user=instance)
+        # Check to make sure the UserProfile migration was run first.
+        # We have to check this because django-guardian creates AnonymousUser
+        # on a post_syncdb signal.  TODO: this will be fixable in Django 1.7 or 1.8
+        # with the post_migrate signal (guardian should use post_migrate and not
+        # post_syncdb).
+        if MigrationHistory.objects.filter(app_name='users').exists():
+            UserProfile.objects.create(user=instance)
 
 
 post_save.connect(create_user_profile, sender=User)
 
 
 def delete_user_profile(sender, instance, **kwargs):
-    instance.userprofile.delete()
+    if hasattr(instance, 'userprofile'):
+        instance.userprofile.delete()
 
 # Delete UserProfile when User is deleted.  We need to do this explicitly
 # because we're monkeypatching the User model (for now).
